Sidebar navigation for settings, notifications, and moderation
